// src/app/enums/SpecialKey.enum.ts
export enum SpecialKey {
  ENTER = 'Enter',
  TAB = 'Tab',
  ARROW_UP = 'ArrowUp',
  ARROW_DOWN = 'ArrowDown',
  ARROW_LEFT = 'ArrowLeft',
  ARROW_RIGHT = 'ArrowRight',
  BACKSPACE = 'Backspace',
  PAGE_DOWN = 'PageDown',
  PAGE_UP = 'PageUp',
  HOME = 'Home',
  END = 'End',
  INSERT = 'Insert',
  DELETE = 'Delete',
  ESCAPE = 'Escape',
  F1 = 'F1',
  F2 = 'F2',
  F3 = 'F3',
  F4 = 'F4',
  F5 = 'F5',
  F6 = 'F6',
  F7 = 'F7',
  F8 = 'F8',
  F9 = 'F9',
  F10 = 'F10',
  F11 = 'F11',
  F12 = 'F12'
}